Docs: Update gerrit guidelines for -1 reviews
The -1 review authority has been moved from all registered users to
users in the "reviewers" category. The reviewers group is for people
who have submitted patches to coreboot.

This change is taking the project back to how it was before 2016, and
is not due to any issues that we're seeing. The reason it was initially
changed was that in 2016, before we required all comments to be resolved
so the patch could be merged, it was easy to overlook comments that
should have been addressed. Now that the process has changed, the -1
right is no longer needed for all users simply to bring attention to
the comment.

The feeling in the leadership meeting was that since it's relatively
easy to get to reviewer status, this should not be an undue burden on
anyone.

Change-Id: I0b7f3dcc80b9122b0f923e6703da73391654d26c
Signed-off-by: Martin Roth <gaumless@gmail.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/81177
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Nicholas Chin <nic.c3.14@gmail.com>