Documentation/framework/azalea: Add mainboard documentation
It's important that people understand that this coreboot port is not
going to be a replacement for the UEFI firmware for most people. This
is an unofficial port with non-production-level code from AMD. I hope
to port from FSP to openSIL in the upcoming months, but the openSIL is
strictly proof-of-concept, getting ready for upcoming platforms.

Signed-off-by: Martin Roth <gaumless@gmail.com>
Change-Id: Idc6fe5e194cf17da1e48ca7be3334682c2fba166
Reviewed-on: https://review.coreboot.org/c/coreboot/+/81994
Reviewed-by: Matt DeVillier <matt.devillier@gmail.com>
Reviewed-by: Daniel Schaefer <dhs@frame.work>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>